Move AccessibilitySetting to model.
